feat(recommendations-graphql-schema): [INFRA-1063] recommendations query
Getting recommendations queries in place to establish what data we want to
expose on newtab surfaces.

This is sticking to a relatively minimal set. New data can always be introduced
to existing APIs. Removing data is more complicated.

Please see associated PR for any discussion with the Fx client team and more
details.