chipper(auth-refactoring): auth refactoring
This is being driven by two motivations. We need to handle
unauthenticated requests for the next route, and we need to
pass auth details into sentry.

Splitting consumer_key from the rest of auth in order to
support consumer_key for unauthenticated requests.

Implementing web session auth as a closure that will support
multiple authentication methods in the future, and to make
it very easy to "do the right thing" WRT sentry tags and
client authentication.

Also implemented more extensive testing for all auth middlewares.

All of this has been end to end tested locally for all existing
routes. If desired, I'd also encourage deploying it to dev.

This should be completely compatible for clients to the previous
API implementation, the only detail that has changed is that there
are now two separate error ids for lookup from sentry errors, one
for WebSessionAuthHandler, and one for ConsumerKeyHandler now that
they are separately handled in preparation for unauthenticated
requests.