feat(recommendations-client): [INFRA-1105] recommendations graphql client
GraphQL client implementation for recommendations, and a corresponding test mock.

There are several properties on both saves and recommendations that we expect to be
approximately equivalent. These were pulled into a set of common mock components

No tests for this yet. API implementation will utilize the mock in its tests.