feat(observability-proto): common observability tools prototypes
Code re-use candidates. Please let me know what you think of these,
and if you have any requests for changes to these tools.

This is an attempt to provide a "batteries included" express server
for use in proxy services, and possibly as a base for GraphQL servers.

I've tried to be pretty comment and test heavy. More extensive testing
for server.ts is incoming, but I wanted to get a round of feedback
before I implement testing for all of the current functionality. I
don't know for sure if all of it is wanted. Please see the comments
and READMEs for more details about what this is and why.