chore(auth-bugs): authentication maintenance, bugfixes
Sorry for the book, trying to document lessons learned. Two fixes here:

1. Remove references to `sess_guid` cookie entirely from service.

This service relies on cookie based auth, and validates cookies before sending requests off to the GraphQL proxy. There are a couple reasons I made this decision:
- This service doesn't live on the same domain as the graphql proxy, so it can't really act as a dumb proxy without introducing synchronization issues between getpocket.com cookies and readitlater.com cookies.
- From my own personal experiences, older cookie based auth flows really only work well when evaluated in a browser window context.

`sess_guid` is also documented as a requirement for WebSession auth in all of the documentation I've read around it, and the php implementation of WebSession suggests it will throw an exception if it is not present. Between this, and the concerns above, validating cookies and giving an explicit error seemed like a better outcome than dealing with opaque errors upstream. However, it seems there is another middleware that will populate `sess_guid` and send a set-cookie header response that evaluates before the WebSession auth handler, so it is not actually a requirement or a necessary part of auth! I still cannot proxy this set-cookie response reliably, so I am just removing references to this cookie entirely. Some other mechanism will have to be used for session tracking if we need this in our metrics.

2. Support `consumer_key` via query params in addition to headers.

This is more speculative, and hasn't had a recorded reproduction, but I have heard a client report that `consumer_key` validation may have failed. I wanted to support client identifiers via headers to allow configuring a client rather than having to include it in every request. `consumer_key` is not actually valid as defined in RFCs (_ is not valid). Many servies support it anyway. I suspect some weird edge case in a load balancer could be dropping it if it attempts to inspect headers. Just getting this in while we're touching adjacent things.

Both of these changes are backwards compatible with deployments in firefox nightly.