feat(dumber-proxy): error handling improvements follow-up (#36)
Continued follow-up for [INFRA-1156] now that error reporting
is working and I'm not guessing. This is a continuation of
error handling efforts started there.

Lots of test catchup.

This is a start of work to transition this service to being a
dumb proxy. More work is required upstream to get the rest of
the way there, but it's a start. This service now does less to
prevent errors proactively, and more to handle errors that result
upstream as a consequence. It has also been modified to forward all
available upstream error details. See the top level README and API
docs for more details. Additional places we want to be a dumber proxy
but cannot yet are marked with TODOs.

WebSessionAuth / Handler has been changed to forward more headers.
Unfortunately the web repo closes the connection immediately if
all errors are forwarded indiscriminately, so this is using an
allow-list for now. This now controls which headers may be forwarded
and primarily acts as a best effort sentry / log tag provider.

The GraphQLClient has been adapted to support forwarding all response
headers from the GraphQL request to the Firefox clients.

other general improvements:

- extended the top level server error handler to support error callbacks
  - **ALL** service errors are now logged via this callback.
- tweaks to server error reporting. Sentry does not support arbitrary extension of express Requests, moving request specific identifying details to request.user.
- Top level GraphQL error handling is now in place for universal errors, and may be extended with route specific error handling as needed.
  - All errors are now handled via GraphQLErrorHandler and top level handlers. Cannot accidentally lose events or logs via missing implementation.
- Improved type safety for graphql-proxy client. Mock implementations now complain when the client is extended without mock follow-up.