fix(proxy-headers): transition header forwarding to block list
There are some issues with the current header forwarding. In particular,
some of the content realated programmatic headers are not being set by express,
and are being set upstream in the graph proxy. These content mis-matches are
resulting in errors in firefox.

This fix transitions the service to blocking header forwarding based on a
block-list, rather than permitting headers based on an allow-list. Hopefully
this is more stable.

Additionally, I did some digging into proxy headers, and found the spec for
hop-by-hop headers, and added all of these to the block-list.