more deployment bugfixes - remove all cdktf array syntax
Remove all array syntax for all of mozillaOpsSourceRule waf rule.

I'm not sure why type checking is failing so badly here, but I think it
might be related to not having cdktf declared as a dependency, and just
implicitly inheriting it from terraform-modules.

Removed every array other than textTransformation (it contains priorities,
which suggests it needs to be an array).